// app/(saas)/usecases/new/types.ts
export interface ClosedFieldOption {
  label: string
  examples?: string[]
  tooltip?: {
    title: string
    shortContent: string
    fullContent?: string
    icon?: string
  }
}